Refactor out product document upload into an action
