+++ exporter/applications/views/goods/add_good_material/views/add.py
@@ -8,8 +8,8 @@ from django.urls import reverse
 from core.auth.views import LoginRequiredMixin
 from core.decorators import expect_status
 
+from exporter.applications.views.goods.common.actions import ProductDocumentAction
 from exporter.core.wizard.views import BaseSessionWizardView
-from exporter.core.helpers import get_document_data
 from exporter.goods.forms.common import (
     ProductControlListEntryForm,
     ProductDescriptionForm,
@@ -27,7 +27,7 @@ from exporter.goods.forms.common import (
     ProductMilitaryUseForm,
 )
 
-from exporter.goods.services import post_material, post_good_documents
+from exporter.goods.services import post_material
 from exporter.applications.services import post_material_good_on_application
 from exporter.applications.views.goods.common.mixins import (
     ApplicationMixin,
@@ -89,31 +89,6 @@ class AddGoodMaterial(
             kwargs["request"] = self.request
         return kwargs
 
-    def has_product_documentation(self):
-        return self.condition_dict[AddGoodMaterialSteps.PRODUCT_DOCUMENT_UPLOAD](self)
-
-    def get_product_document_payload(self):
-        data = self.get_cleaned_data_for_step(AddGoodMaterialSteps.PRODUCT_DOCUMENT_UPLOAD)
-        document = data["product_document"]
-        payload = {
-            **get_document_data(document),
-            "description": data["description"],
-        }
-        return payload
-
-    @expect_status(
-        HTTPStatus.CREATED,
-        "Error with product document when creating material",
-        "Unexpected error adding material to application",
-    )
-    def post_product_documentation(self, good):
-        document_payload = self.get_product_document_payload()
-        return post_good_documents(
-            request=self.request,
-            pk=good["id"],
-            json=document_payload,
-        )
-
     def get_context_data(self, form, **kwargs):
         ctx = super().get_context_data(form, **kwargs)
 
@@ -153,8 +128,8 @@ class AddGoodMaterial(
     def done(self, form_list, form_dict, **kwargs):
         good, _ = self.post_material(form_dict)
         self.good = good["good"]
-        if self.has_product_documentation():
-            self.post_product_documentation(self.good)
+
+        ProductDocumentAction(self).run()
 
         return redirect(self.get_success_url())
